More Updates to Presentation and Service Layer for Librarians
I added the function to view all the books when adding books to a
branch.  Allow the user to create bookCopy when adding a book to the
library that previously had 0.  Also, when a librarian sets a noOfCopies
to 0, it will delete that book from the branch(This can easily be
removed from functionality).  DAO logic and functionality should be all
up to date with needs and wants.  Presentation layer has full core
functionality. Service layer is also functional.  Next I will be adding
test cases.